requirements.txt
alembic==1.14.0
anyio==4.6.2.post1
APScheduler==3.10.4
async-timeout==5.0.1
blinker==1.9.0
cachelib==0.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
exceptiongroup==1.2.2
fake-useragent==1.5.1
Flask==3.0.3
Flask-Caching==2.3.0
Flask-Mail==0.10.0
flask-redis==0.4.0
Flask-SQLAlchemy==3.1.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.6
MarkupSafe==3.0.2
packaging==24.2
pillow==11.0.0
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-telegram-bot==21.7
pytz==2024.2
qrcode==8.0
redis==5.2.0
requests==2.32.3
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.3
Werkzeug==3.1.3
